concord-server: move exception classes to sdk
