+++ server/impl/src/main/java/com/walmartlabs/concord/server/agent/AgentResource.java
@@ -21,8 +21,8 @@ package com.walmartlabs.concord.server.agent;
  */
 
 import com.walmartlabs.concord.server.AgentWorkerUtils;
-import com.walmartlabs.concord.server.boot.validation.ValidationErrorsException;
 import com.walmartlabs.concord.server.sdk.metrics.WithTimer;
+import com.walmartlabs.concord.server.sdk.validation.ValidationErrorsException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.sonatype.siesta.Resource;
