+++ server/impl/src/main/java/com/walmartlabs/concord/server/audit/AuditLogResource.java
@@ -21,7 +21,6 @@ package com.walmartlabs.concord.server.audit;
  */
 
 import com.walmartlabs.concord.server.OffsetDateTimeParam;
-import com.walmartlabs.concord.server.boot.validation.ValidationErrorsException;
 import com.walmartlabs.concord.server.cfg.AuditConfiguration;
 import com.walmartlabs.concord.server.org.OrganizationEntry;
 import com.walmartlabs.concord.server.org.OrganizationManager;
@@ -35,6 +34,7 @@ import com.walmartlabs.concord.server.org.secret.SecretManager;
 import com.walmartlabs.concord.server.org.team.TeamDao;
 import com.walmartlabs.concord.server.sdk.ConcordApplicationException;
 import com.walmartlabs.concord.server.sdk.metrics.WithTimer;
+import com.walmartlabs.concord.server.sdk.validation.ValidationErrorsException;
 import com.walmartlabs.concord.server.security.Roles;
 import com.walmartlabs.concord.server.user.UserDao;
 import io.swagger.v3.oas.annotations.Operation;
