+++ server/impl/src/main/java/com/walmartlabs/concord/server/boot/validation/ValidationErrorsExceptionMapper.java
@@ -31,6 +31,9 @@ package com.walmartlabs.concord.server.boot.validation;
  * =====
  */
 
+import com.walmartlabs.concord.server.sdk.validation.ValidationErrorXO;
+import com.walmartlabs.concord.server.sdk.validation.ValidationErrorsException;
+
 import javax.ws.rs.ext.Provider;
 import java.util.List;
