+++ server/impl/src/main/java/com/walmartlabs/concord/server/repository/listeners/TriggerRefreshListener.java
@@ -21,10 +21,10 @@ package com.walmartlabs.concord.server.repository.listeners;
  */
 
 import com.walmartlabs.concord.process.loader.model.ProcessDefinition;
-import com.walmartlabs.concord.server.boot.validation.ValidationErrorsException;
 import com.walmartlabs.concord.server.org.project.ProjectValidator;
 import com.walmartlabs.concord.server.org.project.RepositoryEntry;
 import com.walmartlabs.concord.server.org.triggers.TriggerManager;
+import com.walmartlabs.concord.server.sdk.validation.ValidationErrorsException;
 import org.jooq.DSLContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
